Incorporate feedback from range-operators
